Create SoccerSim.java
I had a really tough time with both this assignment and the last assignment, but nonetheless I hope that this works, I am going to include the Ball portion of the code in another folder, either way I hope to learn from the notes you make so that I can get full credit on the next assignment! Whether it works or not this time I feel that I have improved even if a little, given that I am not very java or computer science savvy, soon though I will get the hang of it and improve.